move eaw_decompose and eaw_synthesize to new file common/eaw.c
